pyproject.toml
[tool.poetry]
name = "api-tdd"
version = "0.1.0"
description = ""
authors = ["mielesantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
fastapi-pagination = "^0.12.26"
uvicorn = "^0.30.1"
motor = "^3.5.0"
pre-commit = "^3.7.1"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.8"
httpx = "^0.27.0"

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"
asyncio_mode = "auto"
xfail_strict = true
jnuit_family = "xunit2"

[tool.taskipy.tasks]
lint = 'ruff check .&& ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev store/main.py'
precommit = 'poetry run pre-commit install'
test = 'pytest -s -x --cov=app_tdd -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
